Extract Card to be a super generic Component
