+++ ui-kit/Card/index.js
@@ -1,3 +1,3 @@
-import Card from './Card';
+import Card from './Card.styles';
 
 export default Card;
